Factor out Versions class to new lydiff module
This is a first step towards a separation of CLI program and
Python module
